Langchain (#1)
* langchain

* langchain_core

* refactor: update import paths and configuration settings

- Changed import paths from langchain_core to ngerekrut for better organization and clarity.
- Updated analysis_options.yaml to ignore 'prefer_relative_imports' error.
- Removed langchain_core dependency from pubspec.yaml and updated pubspec.lock accordingly.
- Adjusted various files to reflect the new directory structure and improve maintainability.

* chore: update analysis options and refine message metadata handling

- Added directive ordering to analysis_options.yaml to ignore related errors.
- Simplified metadata assignment in ObjectBoxMessageRepository for better readability.
- Adjusted return type in chat_animated_list.dart to remove unnecessary type casting.

---------

+++ lib/flutter_chat_ui/src/chat_animated_list/chat_animated_list.dart
@@ -357,7 +357,7 @@ class _ChatAnimatedListState extends State<ChatAnimatedList>
     if (!ratio.isFinite) {
       return 0.0;
     }
-    return ratio.clamp(0.0, 1.0) as double;
+    return ratio.clamp(0.0, 1.0);
   }
 
   /// If the scroll-to-bottom button should be shown.
